# swagger.yaml
swagger: "2.0"
info:
  title: UrlShortener
  description: Convert long link to short link
  version: 0.0.1
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
  - https
basePath: /api
host: "localhost:8080"
...
responses:
  unauthorized:
    description: Unauthorized
    schema:
      $ref: "swagger/responses/error.v1.yml"
  invalid-request:
    description: Invalid request
    schema:
      $ref: "swagger/responses/error.v1.yml"
...
  server-error:
    description: Internal Server Error
    schema:
      $ref: "swagger/responses/error.v1.yml"

# Paths
paths:
...
  /url:
    put:
      tags:
        - ShortUrl
      summary: Create short URL
      description: Create short URL
      operationId: createShortUrl
      parameters:
        - name: data
          in: body
          schema:
            type: object
            properties:
              url:
                type: string
              days:
                type: integer
              minutes:
                type: integer
              hours:
                type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: "swagger/responses/short_url.yml"
        400:
          $ref: "#/responses/invalid-request"
        # 401:
        #   $ref: "#/responses/unauthorized"

        500:
          $ref: "#/responses/server-error"
    delete:
      tags:
        - ShortUrl
      summary: Delete short URL
      description: Delete short URL
      operationId: deleteShortUrl
      parameters:
        - name: data
          in: body
          schema:
            type: object
            properties:
              short-url:
                type: string
      responses:
        200:
          description: no content
        400:
          $ref: "#/responses/invalid-request"
        # 401:
        #   $ref: "#/responses/unauthorized"

        500:
          $ref: "#/responses/server-error"
...